[Id-Controller] Add template engine and migrate initial banner
The project uses now the razor template engine to generate HTML dynamically.
The tests are removed as well because there is nothing to test now as the
HTML is properly generated. There is only a placeholder test because the
UTest project must not be empty to compile.

The template has been factorized in a way that extra CSS and scripts may
be added when needed.

Change-Id: Iecd1d3956c02fe4431e8ac419a8a3d53dc1a880b
JIRA:IDPS-454