[Id-Controller] Inject JS code inline in the view
The widget is intended to be displayed as soon as possible so we'd like
to reduce external calls (i.e. scripts). This code injects JS code directly
in the template once rendered, so no extra calls are needed.
It implements a caching mechanism as well, so then reducing rendering times.

Change-Id: I47c0ab12069607e5079f4325e4d3fa2d6c7cbfa6