[OTP] Add validation endpoint - 5
This commit adds the validation endpoint so the whole flow can be tested.
Note that there is a Console.Out.WriteLine command to show the OTP code
just generated when development so the flow can be tested without requiring
emails.
The OTP is removed once used (one time use).

Change-Id: Ia84da72345217a8db08e49c36cda41223b26ddb5
JIRA:USRF-12