Introduce request models to use JSON in POST methods
This review adds request models so the data can be easily sent through
JSON from the front-end. This is an laternative to using query string
parameters.

Now the UserAgent string is obtained directly from the User-Agent header
when calling the controller endpoint, so code is cleaner now.

Adapted Event and OTP endpoints, as well as JS code used by the widget.

Change-Id: I2fde0c5007008e9660dcd2b7b27436cb5ac779f3
JIRA:USRF-12