[OTP] Extract code generation logic to a helper and validations
Now there is a helper (service) dedicated to OTP creation so it's easier
to mock and manage in case other systems need it. The code validation has
been extracted to this new class to be aligned with the definition.

The email validation has been moved to the EmailHelper to centralize the
email behaviour there as well. Note that this method validation could have
been done static but better like this to align both interfaces.

Change-Id: If4764b7ccdff349451f8e4f94cec9f9ea882c2a9
JIRA:USRF-12