[OTP] Return token and set cookie header in user validation
When user validates the code, we write a "session" cookie. It is a Guid
for simplicity for now.

By returning this token, the front-end would have something to return
to the parent site which will be stable as long as the cookie is available.

Using the token as cookie session is not secure, will do brainstorming and
follow-up reviews once the best strategy has been decided.

Change-Id: If59e3e88ee729c51a11da933ef9b04554886aab5
JIRA:USRF-29