[OTP] Add email template and view render helper
We need a template engine to generate emails in a fancy way. For that,
we use the Razor template engine already available in asp net core projects.
This template engine is used by the MVP engine out of the box, but we need
to add some code to run in outside a Controller context.

Consequently, this commit adds the following changes:
    - Add helper to render Razor views outside a controller context
    - Add the email template and use it when generating an OTP code

Change-Id: I427de898b2e81242b334ba8add56cbcd8a43b124
JIRA:USRF-31